Trait rusty_p4::app::App

pub trait App: Sync + Send + 'static + Sized {
    type Container: App + Clone;
    type Dependency: Dependencies;
    type Option: AppOption;

    const Name: &'static str;

    fn init<S>(
        dependencies: Self::Dependency, 
        store: &mut S, 
        option: Self::Option
    ) -> Self
    where
        S: AppStore;

    fn from_inner(app: Option<Self::Container>) -> Option<Self>;

#[must_use]    fn run<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

App is the core concept to build a controller, which is for custom logic and function. It can have dependency and option (also called configuration). Using dependency app can get instance of other apps when initializing. So a app is required to be clone + Send + Sync, as it might be held by multiple apps and accessed from different threads.

Associated Types

type Container: App + Clone

To support special function (like optional dependency), some generic container type (like Arc or Option) can also be defined as app. If you create a container type (Option<T>), use this field to specify the target type (T). For regular app, this field should be Self.

type Dependency: Dependencies

Use this field to specify your dependency. It should be a variadic tuple, but variadic tuple is not supported currently. So now, rusty-p4 use crate tuple_list to define dependencies. Use () for no dependency.

type Option: AppOption

Loading content...
Associated Constants

const Name: &'static str

Name your app please.

Loading content...
Required methods

fn init<S>(
dependencies: Self::Dependency,
store: &mut S,
option: Self::Option
) -> Self where
S: AppStore,

This is where app get their dependencies, option then initialize, returns a new instance of app. Container type should not be installed directly, so this method should not be called for container type.

fn from_inner(app: Option<Self::Container>) -> Option<Self>

For container type, use this method to convert from target type to container type (T -> Option<T>). For regular type, simplely return will do. This method is called by rusty-p4 internally.

#[must_use]fn run<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...