Module rusty_p4::proto::p4runtime

Modules

action
Nested message and enum types in Action.

action_profile_action
Nested message and enum types in ActionProfileAction.

action_profile_group
Nested message and enum types in ActionProfileGroup.

digest_entry
Nested message and enum types in DigestEntry.

entity
Nested message and enum types in Entity.

field_match
Nested message and enum types in FieldMatch.

forwarding_pipeline_config
Nested message and enum types in ForwardingPipelineConfig.

get_forwarding_pipeline_config_request
Nested message and enum types in GetForwardingPipelineConfigRequest.

p4_data
Nested message and enum types in P4Data.

p4_runtime_client
Generated client implementations.

packet_replication_engine_entry
Nested message and enum types in PacketReplicationEngineEntry.

set_forwarding_pipeline_config_request
Nested message and enum types in SetForwardingPipelineConfigRequest.

stream_error
Nested message and enum types in StreamError.

stream_message_request
Nested message and enum types in StreamMessageRequest.

stream_message_response
Nested message and enum types in StreamMessageResponse.

table_action
Nested message and enum types in TableAction.

table_entry
Nested message and enum types in TableEntry.

update
Nested message and enum types in Update.

write_request
Nested message and enum types in WriteRequest.

Structs

Action

ActionProfileAction

ActionProfileActionSet

ActionProfileGroup

ActionProfileMember

CapabilitiesRequest

CapabilitiesResponse

CloneSessionEntry
A packet may be cloned by setting the clone_session_id field of PSA ingress/egress output metadata to session_id of a programmed clone session entry. Multiple clones may be created via a single clone session entry by using the replicas message. The clones may be distinguished in the egress using the instance field. The class_of_service field of the clone's egress input metadata will be set to the respective value programmed in the clone session entry. Note that in case of multiple clones, all clones, defined for a clone session, will get the same class of service. The packet_length_bytes field must be set to a non-zero value if the clone packet(s) should be truncated to the given value (in bytes). The packet length is also common to all clones in the clone session. If the packet_length_bytes field is 0, no truncation on the clone(s) will be performed.

CounterData

CounterEntry

DigestEntry

DigestList

DigestListAck
Used by the controller to ack a DigestList message. To avoid flooding the controller, the switch must not generate digest notifications for the same data until a DigestListAck message with the same list_id is received or the ack timeout (ack_timeout_ns field in DigestEntry.Config) expires.

DigestListAckError

DirectCounterEntry

DirectMeterEntry

Entity

Error
Error message used to report a single P4-entity error for a Write RPC.

ExternEntry

FieldMatch
field_match_type ::= exact | ternary | lpm | range | optional

ForwardingPipelineConfig

GetForwardingPipelineConfigRequest

GetForwardingPipelineConfigResponse

IdleTimeoutNotification

Index
An index as a protobuf message. In proto3, fields cannot be optional and there is no difference between an unset integer field and an integer field set to 0. This is inconvenient for reading from P4 array-like structures, such as indirect counters and meters. We need a way to do a wildcard read on these but we cannot use a default zero index value to do so, as zero is a valid index (first entry in the array). We therefore wrap the index in a message.

MasterArbitrationUpdate

MeterConfig
Modeled as RFC 2698: A Two Rate Three Color Marker (trTCM) The trTCM meters a packet stream and marks its packets based on two rates, Peak Information Rate (PIR) and Committed Information Rate (CIR), and their associated burst sizes to be either green, yellow, or red. A packet is marked red if it exceeds the PIR. Otherwise it is marked either yellow or green depending on whether it exceeds or doesn't exceed the CIR.

MeterEntry

MulticastGroupEntry
The (egress_port, instance) pair must be unique for each replica in a given multicast group entry. A packet may be multicast by setting the multicast_group field of PSA ingress output metadata to multicast_group_id of a programmed multicast group entry. The egress_port and instance fields of each replica's egress input metadata will be set to the respective values programmed in the multicast group entry.

P4Data

P4Header

P4HeaderStack

P4HeaderUnion

P4HeaderUnionStack

P4StructLike

P4Varbit

PacketIn
Packet sent from the switch to the controller.

PacketMetadata
Any metadata associated with Packet-IO (controller Packet-In or Packet-Out) needs to be modeled as P4 headers carrying special annotations @controller_header("packet_out") and @controller_header("packet_in") respectively. There can be at most one header each with these annotations. These special headers are captured in P4Info ControllerPacketMetadata.

PacketOut
Packet sent from the controller to the switch.

PacketOutError

PacketReplicationEngineEntry

ReadRequest

ReadResponse

RegisterEntry

Replica
Used for replicas created for cloning and multicasting actions.

Role

SetForwardingPipelineConfigRequest

SetForwardingPipelineConfigResponse

StreamError
Used by the server to asynchronously report errors which occur when processing StreamMessageRequest messages.

StreamMessageRequest

StreamMessageResponse

StreamOtherError

TableAction
table_actions ::= action_specification | action_profile_specification

TableEntry

Uint128

Update

ValueSetEntry

ValueSetMember
A member in a P4 value set. Each member defines a list of matches, which can have different match types.

WriteRequest

WriteResponse

Enums

SdnPort