Trait rusty_p4::service::Service

pub trait Service {
    type Request: ParseRequest + Send;

    const NAME: &'static str;

    fn process(
        &mut self, 
        request: Request<Self::Request>
    ) -> Result<Option<usize>>;
}

Service is used to expose an application to various northbound api. The app implement the trait Service to process request and produce a stream of response. Northbound api might be implemented as multiple different backend. The service is compatiable to tower eco system.

Associated Types

type Request: ParseRequest + Send

Your request type must be deserialized from DefaultRequest, which has:

a path: a Vec.

a action: a String like 'get' or 'set'

parameters: a HashMap<String, String>

Loading content...
Associated Constants

const NAME: &'static str

Loading content...
Required methods

fn process(&mut self, request: Request<Self::Request>) -> Result<Option<usize>>

process your request and send back response via request.respond(), the return value Option<usize> is the size hint (upper bound) of your response stream.

Loading content...